fixes issue described in the chat

+++ app/jobs/account/account_deletion_scheduler_job.rb
@@ -13,6 +13,9 @@ class Account::AccountDeletionSchedulerJob < ApplicationJob
     deletion_days = GlobalConfig.get(deletion_days_config_name)[deletion_days_config_name] || 3
     total_no_days = intemediatry_days + no_days + deletion_days.to_i
 
+    auto_delete_inactive_account_name = 'AUTO_DELETE_INACTIVE_ACCOUNT'
+    auto_delete_inactive_account = GlobalConfig.get(auto_delete_inactive_account_name)[auto_delete_inactive_account_name] || false
+
     Account.where(deletion_email_reminder: :second_reminder).each do |account|
       subscription = account.account_billing_subscriptions.where(cancelled_at: nil)&.last
 
@@ -21,15 +24,19 @@ class Account::AccountDeletionSchedulerJob < ApplicationJob
       users = account.users.where('last_sign_in_at > ? ', total_no_days.days.ago)
 
       if users.present?
-        account.update(deletion_email_reminder: nil)
+        if account.users.where('last_sign_in_at > ? ', (intemediatry_days + no_days).days.ago).present?
+           account.update(deletion_email_reminder: nil)
+        end
       else
         user = account.account_users.where(inviter_id: nil).last&.user
         next if user.blank?
         next unless user.created_at < no_days.days.ago
 
         AdministratorNotifications::AccountMailer.account_deletion(account).deliver_now
-        account.users.destroy_all
-        account.destroy
+        if auto_delete_inactive_account == 'true'
+          account.users.destroy_all
+          account.destroy
+        end
       end
     end
   end
